requirements.txt
# Data Processing
pandas==2.2.0
numpy==1.26.3

# BigQuery
google-cloud-bigquery==3.17.2
google-auth==2.27.0
db-dtypes==1.2.0
protobuf==4.25.3


# dbt
dbt-core==1.7.7
dbt-bigquery==1.7.4

# WhatsApp (we'll add this later)
# twilio==9.0.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
